Propose custom OSVersion type and PolicyManager.
First, a new OSVersion type, specific for Nudge.
It can be initialized from both the OperatingSystemVersion type Apple provides
or by parsing a string. Having a custom type has several benefits and guarantees.
For example, configuration values (usually strings) will be parsed consistently.
The type has additonal extensions making it comparable,
equatable and convertible to a custom string.

My second change is a bit more ambitious, so I left the implementation
minimal, with the intention to grow it if accepted.
I created a "PolicyManager" class (naming things is hard), passed
to the ContentView as an EnvironmentObject. The policy manager is intended to
separate the UI from all the actions Nudge would take, like loading
preferences, getting the OS version, launching System Preferences and so
on. I'd like to refactor all the standalone helpers/global values into
this class and give it more structure.