make decisions by evaluating policy
Picks the required update from configuration and compares the system
against it. Returns an action as a result.

I refactored the string typed version in config to be a OSVersion, which
allows everything to be easily comparable. In the process I ended up
cutting some corners with the current globals. The changes are having a
domino effect where it's hard to change one thing without changing a
dozen other helpers in osutils/nudge.

Don't merge this yet/leaving for discussion before we decide to do
bigger changes or take a different approach.