Minor updates
Very minor updates for now.  Updated the two .json + localizable strings and the README. I think some of the verbiage can definitely be made better in French but it would require doing more than a 1:1 translation. If you are comfortable with that let me know and I can make another more significant update (and tell you in the PR what the new stuff means)