Move to a single nudgeRefreshCycleTimer for both standard/simple mode
When refactoring the standardMode UI to LeftSide/RightSide in v1.0.0, multiple timers and repeated code was used for ensure the UI would stay in sync together. This is because I did not know about ObservableObject.

Now that a lot of the UI variables have been moved to OO, Nudge no longer needs multiple timer controllers running together to update the UI.